rr-config: cast mode id to uint32_t
In some places it’s used as an unsigned int, in other cases it’s signed,
so this just casts the -1 in a ternary expression to uint32_t, which
makes no difference.

+++ libgnome-desktop/gnome-rr-config.c
@@ -1168,7 +1168,7 @@ crtc_assignment_apply (CrtcAssignment *assign, gboolean persistent, GError **err
 	g_variant_builder_add (&crtc_builder, "(uiiiuaua{sv})",
 			       gnome_rr_crtc_get_id (crtc),
 			       info->mode ?
-			       gnome_rr_mode_get_id (info->mode) : -1,
+			       gnome_rr_mode_get_id (info->mode) : (guint32) -1,
 			       info->x,
 			       info->y,
 			       rotation_to_transform (info->rotation),
